requirements.txt
imapclient
python-dotenv
beautifulsoup4
mail-parser
google-api-python-client
google-auth
google-auth-oauthlib